upd examples: move multisig config to separate directory & move buildAndLogMultisigTransaction to separate file
